add shell script; add agent conf; abstract aworld agent loop
